+++ ptz-visca.hpp
@@ -185,13 +185,13 @@ public:
 	virtual OBSData get_config() = 0;
 
 	void cmd_get_camera_info();
-	void pantilt(double pan, double tilt);
+	void pantilt(int pan, int tilt);
 	void pantilt_rel(int pan, int tilt);
 	void pantilt_stop();
 	void pantilt_home();
 	void zoom_stop();
-	void zoom_tele(double speed);
-	void zoom_wide(double speed);
+	void zoom_tele(int speed);
+	void zoom_wide(int speed);
 	void memory_reset(int i);
 	void memory_set(int i);
 	void memory_recall(int i);
